102-fibonacci

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -5,19 +5,19 @@
  */
 int main(void)
 {
-	int fibonacci[50];
+	long fibonacci[50];
 	int i;
 
 	fibonacci[0] = 1;
 	fibonacci[1] = 2;
 
-	printf("%d, %d", fibonacci[0], fibonacci[1]);
+	printf("%ld, %ld", fibonacci[0], fibonacci[1]);
 
 	for (i = 2; i < 50; i++)
 	{
 		fibonacci[i] = fibonacci[i - 1] + fibonacci[i - 2];
 
-		printf(", %d", fibonacci[i]);
+		printf(", %ld", fibonacci[i]);
 	}
 	printf("\n");
